changed webapp structures
