+++ src/webapp/package.json
@@ -10,9 +10,9 @@
     "dev": "vue-cli-service build --watch --mode development --dest builds/dev"
   },
   "dependencies": {
+    "axios": "^0.18.0",
     "vue": "^2.5.17",
     "vue-material-design-icons": "^1.6.0",
-    "vue-resource": "^1.5.1",
     "vue-router": "^3.0.1",
     "vuetify": "^1.0.19",
     "vuex": "^3.0.1"
