+++ src/webapp/src/api/user.js
@@ -1,0 +1,7 @@
+import config from '../../../config/dev'
+
+export default {
+  // 
+  // 
+  // 
+}
