+++ src/webapp/src/store/index.js
@@ -1,0 +1,14 @@
+import Vue from 'vue'
+import Vuex from 'vuex'
+import user from './modules/user'
+
+Vue.use(Vuex)
+
+const debug = process.env.NODE_ENV !== 'production'
+
+export default new Vuex.Store({
+  modules: {
+    user
+  },
+  strict: debug
+})
